+++ front/src/actions/dashboard/boxes/weather.js
@@ -6,7 +6,36 @@ import get from 'get-value';
 
 const BOX_KEY = 'Weather';
 
-function createActions(store) {
+const translateWeatherToFeIcon = (weather) => {
+  if (weather.search('snow') !== -1) {
+    return 'fe-cloud-snow';
+  }
+  if (weather.search('rain') !== -1) {
+    return 'fe-cloud-rain';
+  }
+  if (weather.search('clear') !== -1) {
+    return 'fe-sun';
+  }
+  if (weather.search('cloud') !== -1) {
+    return 'fe-cloud';
+  }
+  if (weather.search('fog') !== -1) {
+    return 'fe-cloud';
+  }
+  if (weather.search('sleet') !== -1) {
+    return 'fe-cloud-drizzle';
+  }
+  if (weather.search('wind') !== -1) {
+    return 'fe-wind';
+  }
+  if (weather.search('night') !== -1) {
+    return 'fe-moon';
+  }
+  return 'fe-question';
+};
+
+
+const createActions = (store) => {
   const boxActions = createBoxActions(store);
 
   const actions = {
@@ -17,6 +46,11 @@ function createActions(store) {
         weather.datetime_beautiful = dayjs(weather.datetime)
           .locale(state.user.language)
           .format('dddd DD MMMM');
+          weather.temperature = weather.temperature.toFixed(2);
+        weather.weather = translateWeatherToFeIcon(weather.weather);
+        weather.hours.map(day => {
+          day.weather = translateWeatherToFeIcon(day.weather);
+        })
         boxActions.mergeBoxData(state, BOX_KEY, x, y, {
           weather
         });
